semihost: Add support for ARC targets
Qemu for ARC provides semihosting support in a completely different
way than arm or risc-v. Its calls are effectively POSIX calls, which
makes things very easy to support under picolibc. It was missing any
way to get the errno value, but I provided a patch which was merged
upstream to add this.

Signed-off-by: Keith Packard <keithp@keithp.com>