scripts: Update arc scripts to enable tests
With the addition of semihost and picocrt support, the picolibc tests
can now be run on a subset of the supported arc targets -- arcem,
archs and em. Change the do-arc-configure script to build the library
for those targets with tests enabled, adjust the cross compilation
control file to pass suitable compiler flags and add a script to run
the resulting tests under qemu-system-arc.

Signed-off-by: Keith Packard <keithp@keithp.com>