Support per-target linker scripts for tests
qemu for arm doesn't have a single machine which supports all of the
desired cpu targets, instead we need to use a range of them with
different memory configurations.

To make this work with a multilib build, add support for cross
compilation properties for each qemu machine that are mapped to the
relevant gcc targets.

Signed-off-by: Keith Packard <keithp@keithp.com>