scripts: Use readelf to select qemu params in run-arm
This allows us to delete all of the other run scripts for arm targets
as the filename is no longer significant in qemu configuration.

This also adds support for all of the remaining targets including
v5te, v6m and v8.

The v6m code is run using the microbit model which has only 256kB of
flash and 16kB of RAM. Programs larger than this are skipped.

Signed-off-by: Keith Packard <keithp@keithp.com>