Place GOT entries in their own segment
In -static mode, when doing relocations for code built with -fpic
which use the GOT, it appears that the linker needs to know which
segment will contain the addresses. Create a separate .got segment so
that the PLT entries start at the begining of that segment.  This
appears to get the addressing correct on m68k which has a couple of
GOT/PLT entries in libgcc.

When code is built without -fpic (as it should be), this shouldn't
matter.

Signed-off-by: Keith Packard <keithp@keithp.com>

fixup