libm/test: Use 'volatile' to prevent unintended sNaN -> qNaN
Running an sNaN value through any computation will raise FE_INVALID
and convert it to qNaN. If that happens before we manage to get the
value to the relevant math function, then the tests will generate
incorrect answers. On m68k, this appears to happen when doing a
release build (not -Os) in the double->float conversion path somewhere.

Prevent this by flagging all of the argument variables as 'volatile'.

Signed-off-by: Keith Packard <keithp@keithp.com>