.github/Dockerfile*: clean apt-get immediately
The docker build cache is either a complete
hit or a complete miss for the RUN statements
in the dockerfiles.

Since the build cache is binary, put everything
in a single RUN statement to avoid creating extra
layers, and clean up the list files
that apt-get update downloads.

This reduces the image sizes from 4.22GB to 4.2GB,
and from 6.03GB to 5.97GB.

Signed-off-by: Peter A. Jonsson <jonsson.peter.a@gmail.com>