CI: increase Docker compression level
Increasing the compression level from 19
to 20 gives significant space savings
without increasing docker load times.
Increase the compression level to the max
since that will benefit every download
performed by the CI.

Compression level is documented, but it
does currently not have any effect
for the docker exporter. If level 22
turns out to take prohibitively long
when the parameter is no longer ignored,
this setting can be dialed back to 20.

Signed-off-by: Peter A. Jonsson <jonsson.peter.a@gmail.com>