libc/tinystdio: Replace %/ with fancy non-div using code on some targets
Provides custom divide+modulus code for use by printf on targets
without hardware dividers. This custom code avoids requiring soft
divide code, which can be useful if applications aren't otherwise
using division of the relevant size.

For instance, arm targets do not have 64-bit division needed for
decimal output when io-long-long is enabled.

The custom code uses the multiplication-by-reciprocal technique built
out of shifts and adds to approximate the division and then adjusts to
provide an exact quotient and remainder.

This needs to be configured per-target but there isn't any convenient
general-purpose tests available during the build. For now, there are
examples which use architecture-specific preprocessor symbols on Arc,
ARM and Risc-V to use the custom code on targets known to lack
hardware division. For other architectures, the custom code will be
enabled for 64-bit conversions on targets with sizeof(long) < 8.

This code is enabled with the printf-small-ultoa option, which is
'false' by default.

Signed-off-by: Keith Packard <keithp@keithp.com>