libc/tinystdio: Support the 'minimal-io-long-long' option
This option controls long long support in the minimal printf/scanf
code separately from the io-long-long option, which now only controls
long long support in the integer printf/scanf variant.

This required rather significant refactoring of all of the option
support in both printf and scanf. The result is that the 'full'
printf/scanf versions include almost everything, except for long long
and %b support, which remains optional.

The minimal version will never include %b support or positional args,
so it's a bit smaller when those are enabled for the other variants.

Signed-off-by: Keith Packard <keithp@keithp.com>