meson.build: get -mhwmult selection from GCC
Query GCC for the snippet that selects the library
corresponding to the -mhwmult parameter, and
splice that into picolibc.specs.

This makes the tests not link on msp430 since
they are linked with -nostartfiles, so add -lmul_none
in the libgcc config option in scripts since that
is not propagated outside picolibc.

Signed-off-by: Peter A. Jonsson <jonsson.peter.a@gmail.com>