scripts: Silence warning about .got table when linking on aarch64
The .got table gets marked as writable but doesn't need to be. We want
to stick it in read-only memory along with executable bits so the
linker ends up warning about the unified segment being marked
RWX. Silence the linker warning as I can't figure out how to fix it
for real.

Signed-off-by: Keith Packard <keithp@keithp.com>