_FORTIFY_SOURCE only works while building tests
All of the _FORTIFY_SOURCE bits are disabled while building the
library due to the check for _LIBC in features.h. This makes it much
easier to implement the library functionality underneath the fortify
source code.

This means that the fortify-source option can only affect
tests. Restructure how this option is used to have it set the
test_c_args value and ensure that the library's c_args always include
-U_FORTIFY_SOURCE in case the local build environment (as on ubuntu)
always includes this.

Finally, set fortify-source to '2' by default so that we catch faults
in tests.

Signed-off-by: Keith Packard <keithp@keithp.com>