picolibc.ld.in: use separate sections for .rodata and .data.rel.ro
Having these be separate makes it easier to look at binary and see which
sections contribute to total binary size. Additionally, it would allow for
a correct PT_GNU_RELRO if picolibc.ld were to grow PIC support in the
future.

This should not have any effect on the final binary layout, it only adds
two more ELF section headers.