libc/time: Eliminate optimization in tzset comparing previous value
Checking to see if tzset is called with the same value as the last
time requires that the previous value be stored in memory. As that
string can be quite long, that either requires a large static buffer
or the use of malloc.

Removing this optimization from the library and expecting applications
to only call tzset when required saves all of that cost.

Signed-off-by: Keith Packard <keithp@keithp.com>