qsort: allow swapping intptr_t-sized elements
This should improve performance for 64-bit architectures since we now
copy 64 bits at a type instead of 32 and is in fact required for CHERI.

This change was pulled in from CheriBSD: For CHERI, if the array contains
pointers we have to swap pointer-sized elements at a time since partial
writes of a CHERI capability invalidate the validity bits. This change
could have also replaced `long` with `intptr_t`, but having both cases
ensures that if we can't swap a pointer-size element, we use int instead
to avoid falling back fall back to int to avoid using the byte copy loop
for smaller elements.

Should have no code generation difference for existing 32-bit
architectures since the compiler can elide the unreachable branch.