tinystdio: Check the *address* of weak stdin/stdout
__bufio_get wants to flush stdout when reading from stdin, but only if
both stdin and stdout are actually being used by the application. The
hack for this is to declare them as 'weak' symbols; if not used
elsewhere in the application, they'll end up with address zero.

We detect that case by comparing the *address* of the symbols against
NULL, not the value.

Replace the value check for stdout with a check of the address and add
an address check for stdin before fetching that.

Signed-off-by: Keith Packard <keithp@keithp.com>