math/riscv: Providing definitions of exceptions flags for softfloat
Part 7.6 point 5 of ISO/IEC 9899:1999 standard requires the modified macros to exist as long as functions in 7.6.2 are provided (these functions are provided also in fenv-softfloat.h not only fenv-fp.h).
Current Picolibc only defines the macros when extensions ("f", "Zfinx", "Zdinx") are provided.
So we need to adjust these macros to be defined whenever 7.6.2 functions are provided.