tinystdio: Add infrastructure to handle fflush(NULL)
C17 defines fflush(NULL) to flush all currently open files. Tinystdio
doesn't generally keep track of every file that has been opened, and
the overhead for doing so is pretty substantial, so we really don't want to
impose that on every application calling fflush.

Resolve this by making calls to fflush with the constant NULL do all
of this work while calls with a non-constant NULL do not. This is
clearly a compromise designed to allow access to the 'flush
everything' functionality without imposing the overhead on all
applications.

This is done by converting fflush into a macro which detects a
constant NULL and invokes _fflush_null. Other fflush uses will pass
through to the original implementation, now called _fflush_nonnull.

Within the same file as _fflush_null, there are two functions,
_fflush_register_real and _fflush_unregister_real which manage a list
of FILE pointers. Those are then wrapped in functions _fflush_register
and _fflush_unregister which refer to their _real analogs using weak
pointers.

If _fflush_null is never referenced, then the file containing it and
the _real functions will not get linked causing the non-_real
functions to do nothing.

This is not completely spec-compliant, but should work in cases where
users actually do want to use fflush(NULL).

Signed-off-by: Keith Packard <keithp@keithp.com>