tinystdio:vfprintf() does not handle the '%n' format specifier.
%n: It assigns a variable the count of the number of characters used in the
print statement before the occurence of '%n'

before this change '%n' format specifier was not handled in tinystdio.
Therefore, they were treated as normal characters resulting in wrong character
count when used.

Add a new meson option, -Dprintf-percent-n, which controls whether this feature
is enabled or not. This is false by default.

Signed-off-by: Ahmed Shehab <shehab@synopsys.com>