tinystdio:fixing hexadecimal string conversion
Perform the hex adjustment of the digit position exponent value before
parsing the explicit exponent to make sure that the exp is calulated
correctly when there isn't any hex exponent

The hexadecimal string with decimal points was handled in a wrong way when it
was passed to these functions strod(), strtof() and strtold().
The exp variable was only handled when the exp_match is equal to 'p' or 'e',
so when any other hexadecimal string with no e or p format specifiers is passed,
wrong results returned. therefore the change was made so the exp is handled
correctly in all cases not only when p or e are passed as format specifiers.

Added several new testing values in the file test-strtod.c to test the
mentioned case above.

Reported-by: Hana Ashour <hanaa@synopsys.com>
Signed-off-by: Keith Packard <keithp@keithp.com>