mallocr.c: add an implementation of posix_memalign().
The two allocators mallocr.c and nano-mallocr.c mostly support the
same set of functions, but posix_memalign is an exception:
nano-mallocr has an implementation of it, but mallocr doesn't. This
means the available API changes depending on whether you asked for a
fast or a small memory allocator.

mallocr does support aligned memory allocation, via other APIs. So
posix_memalign can be implemented as a wrapper on the existing API,
just as it is in nano-mallocr.