tinystdio: Correct printf %a round-to-even with zero precision
The missing MSB in the significand caused the 'odd' check to fail,
thus failing to round 0x1.8 to 0x2. Insert the implicit MSB into the
significand so that the existing round-to-even code worked correctly.

That required moving the NAN/INF checks before adding the MSB.

Doing this insertion removed the need for the special case for the
leading digit of the result, along with the special case for rounding
that digit as the significand now includes the MSB.

Reported-by: Ahmed Shehab <shehab@synopsys.com>
Signed-off-by: Keith Packard <keithp@keithp.com>