vfscanf: Fix pointer step for `wchar_t` specifers `l` followed by `c` or `s` or `[`
Part 7.19.6.2 of C standard states that `fscanf` that a following `c, s, or [`, to `l`, applies to an argument with type pointer to `wchar_t`.

`WIDE_CHARS` guard is removed  in some occurrences as it prevents the right flow for the function which is taking `wchar_t` pointer step whenever flag `FL_LONG` is set within flags variable.

Signed-off-by: Abdallah Abdelhafeez (Abdallahs70)