tinystdio: Fix vfprintf to not cap the max precision to print a hex format
printing using %.15a or %.15A formats give wrong results because dtox_engine
handles max of 13 figures after decimal point in hex format.
and it carried out that maximum precision to vfprintf function which max out the
number of digits that get printed so if the format given is more than %.13a or
%.13A it will still print 13 digits after the decimal place instead of just adding
zeros after 13 digits which is not the correct behavior for vfprintf.

After this change
vfprintf handles more than 13 figures after the decimal point in hex format.

Signed-off-by: Ahmed Shehab <shehab@synopsys.com>