tinystdio: Support %ls/%lc in scanf, %s/%c in wscanf
Somewhat like printf, scanf needs to deal with mixed-representation
characters in input and output which require using the multi-byte APIs
to translate between wide characters and multi-byte values. So, when those
are available, use them.

To support wscanf '[' conversions, the bracket parsing code was
replaced with a (much) slower implementation that walks over the
format string for each input character to avoid needing to allocate
memory for a more efficient representation.

Signed-off-by: Keith Packard <keithp@keithp.com>